requirements.txt
streamlit
opencv-python-headless
pytesseract
pdf2image
Pillow
python-dotenv
google-generativeai
